refactor(models): move geometry helpers (_geometry_from_chs_info, _normalize_pos) to models.util
