+++ test/unit_tests/models/test_foundation_models.py
@@ -1124,7 +1124,6 @@ def test_codebrain_return_features():
 from braindecode.models import BrainOmni, BrainTokenizer  # noqa: E402
 from braindecode.models.base import EEGModuleMixin  # noqa: E402
 from braindecode.models.brainomni import (  # noqa: E402
-    _geometry_from_chs_info,
     _ResidualVQ,
     _SEANetDecoder,
     _SEANetEncoder,
@@ -1132,6 +1131,7 @@ from braindecode.models.brainomni import (  # noqa: E402
     _SpatialTemporalBlock,
     _TokenizerEncoder,
 )
+from braindecode.models.util import _geometry_from_chs_info  # noqa: E402
 
 # Shared small-model config (keeps every BrainOmni/BrainTokenizer build fast).
 _BRAINOMNI_KW = dict(
